# Implemented PXB-2710 - Predict memory for --prepare