--- a/storage/innobase/include/buf0buf.ic
+++ b/storage/innobase/include/buf0buf.ic
@@ -1115,6 +1115,23 @@ static inline ulint buf_pool_size_align(ulint size) {
   }
 }
 
+#ifdef XTRABACKUP
+/** Calculate aligned buffer pool size based on srv_buf_pool_chunk_unit,
+if needed we round it down.
+@param[in]	size	size in bytes
+@return	aligned size */
+static inline ulint buf_pool_size_align_down(ulint size) {
+  const ulint m = srv_buf_pool_instances * srv_buf_pool_chunk_unit;
+  size = std::max(size, srv_buf_pool_min_size);
+
+  if (size % m == 0) {
+    return (size);
+  } else {
+    return ((size / m - 1) * m);
+  }
+}
+#endif
+
 /** Return how many more pages must be added to the withdraw list to reach the
 withdraw target of the currently ongoing buffer pool resize.
 @param[in]      buf_pool        buffer pool instance