--- a/storage/innobase/xtrabackup/src/utils.cc
+++ b/storage/innobase/xtrabackup/src/utils.cc
@@ -19,6 +19,13 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA
 #include <my_default.h>
 #include <mysqld.h>
 
+#ifdef __APPLE__
+#include <mach/mach_host.h>
+#include <sys/sysctl.h>
+#else
+#include <proc/sysinfo.h>
+#endif
+
 #include "common.h"
 #include "msg.h"
 #include "xtrabackup.h"
@@ -99,6 +106,45 @@ unsigned long get_version_number(std::string version_str) {
     version = stoi(version_str.substr(minor_p + 1));
   return major * 10000 + minor * 100 + version;
 }
+
+#ifdef __APPLE__
+unsigned long host_total_memory() {
+  unsigned long total_mem = sysconf(_SC_PHYS_PAGES) * sysconf(_SC_PAGESIZE);
+  return total_mem;
+}
+
+unsigned long host_free_memory() {
+  unsigned long total_mem = host_total_memory();
+  int64_t used_mem;
+  vm_size_t page_size;
+  mach_msg_type_number_t count;
+  vm_statistics_data_t vm_stats;
+
+  // Get used memory
+  mach_port_t host = mach_host_self();
+  count = sizeof(vm_stats) / sizeof(natural_t);
+  if (KERN_SUCCESS == host_page_size(host, &page_size) &&
+      KERN_SUCCESS ==
+          host_statistics(host, HOST_VM_INFO, (host_info_t)&vm_stats, &count)) {
+    used_mem = ((int64_t)vm_stats.active_count + (int64_t)vm_stats.wire_count) *
+               (int64_t)page_size;
+
+    ut_a(total_mem >= (unsigned long)used_mem);
+    return total_mem - (unsigned long)used_mem;
+  }
+  return 0;
+}
+#else
+unsigned long host_total_memory() {
+  meminfo();
+  return kb_main_total * 1024;
+}
+
+unsigned long host_free_memory() {
+  meminfo();
+  return kb_main_available * 1024;
+}
+#endif
 
 }  // namespace utils
 }  // namespace xtrabackup