https://jira.percona.com/browse/PXB-2710

Description:
Implemented a new functionality to predict memory required to --prepare
a backup.
The logic inherits the calculation done at recv_add_to_hash_table and
populate a new struct that controls the memory will be required.
This information is then saved to xtrabackup_checkpoints file and read
at --prepare to predict the memory.
Memory requirements is split into two buckets:
  1. Memory required for recv_add_to_hash_table.
  2. Memory required for free frames.

A new variable has been introduced --use-free-memory-pc (defatul 50)
that controls the amount of free memory we will be able to use at
--prepare.