@satya-bodapati 2022-06-16
> @satya-bodapati I have done a test using 24 sysbench tables with 500000 records in each:
> 
> ```
> sysbench --db-driver=mysql --db-ps-mode=disable --mysql-host=${WRITE_IP} --mysql-port=${WRITE_PORT} --mysql-user=sysbench --mysql-password=sysbench --table_size=500000 --tables=24 --threads=4  /usr/share/sysbench/oltp_read_write.lua prepare
> ```
> 
> Took a backup with --throttle so it will take longer. It resulted in:
> 
> ```
> cat backup/xtrabackup_checkpoints | grep redo
> redo_memory = 12622946976 # ~11.75G for hash records
> redo_frames = 67240 # ~1G for InnoDB Pages
> ```
> 
> ### With --use-free-memory-pct=50 on a server with (60G free memory)
> 
> ```
> 2022-06-16T09:19:49.145167-03:00 0 [Note] [MY-011825] [Xtrabackup] Got prediction from backup: 12622946976 bytes for parsing and 67240 frames. Requesting 13824425984 for --prepare.
> 2022-06-16T09:19:49.145172-03:00 0 [Note] [MY-011825] [Xtrabackup] Setting buffer pool to: 13824425984
> 2022-06-16T09:19:49.145670-03:00 0 [Note] [MY-011825] [Xtrabackup] Using 13824425984 bytes for buffer pool (set by --use-memory parameter)
> 2022-06-16T09:19:49.146953-03:00 0 [Note] [MY-012955] [InnoDB] Initializing buffer pool, total size = 12.875000G, instances = 1, chunk size =128.000000M
> . . .
> 2022-06-16T09:21:03.674213-03:00 0 [Note] [MY-011825] [Xtrabackup] completed OK!
> 
> real	1m14,757s
> user	1m42,042s
> sys	0m3,880s
> ```
> 
> ### Without --use-free-memory-pct
> 
> ```
> 2022-06-16T09:22:47.878392-03:00 0 [Note] [MY-011825] [Xtrabackup] Using 134217728 bytes for buffer pool (set by --use-memory parameter)
> 2022-06-16T09:22:47.879657-03:00 0 [Note] [MY-012955] [InnoDB] Initializing buffer pool, total size = 128.000000M, instances = 1, chunk size =128.000000M
> . . . 
> 2022-06-16T09:29:24.465986-03:00 0 [Note] [MY-011825] [Xtrabackup] completed OK!
> 
> real	6m36,778s
> user	1m53,152s
> sys	0m14,241s
> ```
> 
> So an improvement from 396 seconds to 74 seconds - a difference of ~535% (about **5x faster** )

Those are astonishing numbers!! Great improvement:clap: